chore(c4container): refactors plantuml module.
