+++ server/pkg/rendering/c4container/compression/blocksplitter.go
