+++ server/pkg/rendering/c4container/compression/gzip_container.go
