+++ server/pkg/rendering/c4container/compression/hash.go
