+++ server/pkg/rendering/c4container/compression/katajainen.go
