+++ server/pkg/rendering/c4container/compression/squeeze.go
