+++ server/pkg/rendering/c4container/compression/tree.go
