+++ server/pkg/rendering/c4container/compression/util.go
