+++ server/pkg/rendering/c4container/plantuml_test.go
@@ -1,4 +1,4 @@
-package plantuml
+package c4container
 
 import (
 	"context"
